Refactor(dataSync): Brake out the sync calls to a controller.
