+++ app/src/main/graphql/com.m.services.dataSync/addTaskSubscription.graphql
