+++ app/src/main/graphql/com.m.services.dataSync/deleteTask.graphql
