+++ app/src/main/graphql/com.m.services.dataSync/deleteTaskSubscription.graphql
