+++ app/src/main/graphql/com.m.services.dataSync/getAllTasks.graphql
