+++ app/src/main/graphql/com.m.services.dataSync/schema.json
